ycotd-python-queue/requirements.txt
python-dotenv
requests
tenacity
bullmq
